pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flowstate-cli"
version = "1.2.2"
description = "Command-line productivity tool with task management and Pomodoro timers"
authors = [{name = "FlowState", email = "sundancegithub@gmail.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "plyer>=2.1.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "sqlalchemy>=2.0.0",
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0"
]

[project.scripts]
flowstate = "flowstate_cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["flowstate_cli*"]
